Move `SensitiveString` and account/user data handling to `synctools`
