+++ core/src/main/kotlin/at/bitfire/davdroid/network/NextcloudLoginFlow.kt
@@ -8,7 +8,7 @@ import androidx.annotation.VisibleForTesting
 import at.bitfire.dav4jvm.ktor.exception.HttpException
 import at.bitfire.davdroid.settings.Credentials
 import at.bitfire.davdroid.ui.setup.LoginInfo
-import at.bitfire.davdroid.util.SensitiveString.Companion.toSensitiveString
+import at.bitfire.synctools.util.SensitiveString.Companion.toSensitiveString
 import at.bitfire.davdroid.util.withTrailingSlash
 import at.bitfire.vcard4android.GroupMethod
 import io.ktor.client.HttpClient
