+++ gradle/libs.versions.toml
@@ -20,7 +20,7 @@ androidx-test-junit = "1.3.0"
 androidx-work = "2.11.2"
 bitfire-cert4android = "b8e2d33f09"
 bitfire-dav4jvm = "1f2f5a7793"
-bitfire-synctools = "06f2de81e1"
+bitfire-synctools = "a546290e96"
 compose-accompanist = "0.37.3"
 compose-bom = "2026.05.00"
 conscrypt = "2.5.3"
